+++ src/App.jsx
@@ -1,27 +1,31 @@
-import { useImmer } from "use-immer";
-import { placeholderAsset } from "./placeholderAsset.ts";
-import Preview from "./components/Preview";
-import Form from "./components/Form";
-import Footer from "./components/Footer";
-import Header from "./components/Header";
+import { useImmer } from 'use-immer';
+import { AssetContext } from 'src/services/AssetContext.js';
+import { placeholderAsset } from 'src/placeholderAsset.ts';
+import Header from 'src/components/Header';
+import Preview from 'src/components/Preview';
+import Form from 'src/components/Form';
+import Footer from 'src/components/Footer';
 
 function App() {
-	const [asset, updateAsset] = useImmer(placeholderAsset);
+    const [asset, updateAsset] = useImmer(placeholderAsset);
 
-	return (
-		<div className="wrapper">
-			<div className="main-container">
-				<Header />
+    return (
+        <div className="wrapper">
+            <div className="main-container">
+                <Header />
 
-				<main>
-					<Preview asset={asset} />
-					<Form updateAsset={updateAsset} />
-				</main>
-			</div>
+                <main>
+                    <AssetContext.Provider value={asset}>
+                        <Preview />
+                    </AssetContext.Provider>
 
-			<Footer />
-		</div>
-	);
+                    <Form updateAsset={updateAsset} />
+                </main>
+            </div>
+
+            <Footer />
+        </div>
+    );
 }
 
 export default App;
