--- src/components/Form.scss
@@ -1,182 +1,0 @@
-.form {
-	border: solid var(--black) 2px;
-	padding: 30px;
-	position: relative;
-
-	ul {
-		list-style-type: none;
-		display: flex;
-		gap: 15px;
-		margin-bottom: 20px;
-
-		li {
-			font-family: var(--ui);
-			font-weight: 700;
-			font-size: 24px;
-			color: var(--grey);
-
-			&:hover {
-				cursor: pointer;
-			}
-
-			&.active {
-				color: var(--black);
-				transform: translateX(0%);
-			}
-		}
-	}
-
-	form {
-		fieldset {
-			border: none;
-			display: flex;
-			flex-direction: column;
-			gap: 20px;
-
-			&:not(.visible) {
-				display: none;
-			}
-		}
-
-		.fieldgroup {
-			display: flex;
-			align-items: start;
-			flex-wrap: wrap;
-			gap: 20px;
-		}
-
-		label {
-			display: block;
-			font-family: var(--ui);
-			font-weight: 600;
-			font-size: 16px;
-			margin-bottom: 6px;
-		}
-
-		input[type="text"],
-		select,
-		textarea,
-		input[type="range"],
-		input[type="file"],
-		label {
-			width: 200px;
-		}
-
-		input[type="checkbox"] + input[type="text"] {
-			margin-left: 10px;
-			width: 170px;
-		}
-
-		input[type="checkbox"]:not(:checked) + input[type="text"],
-		input[type="checkbox"]:not(:checked) + input[type="number"] {
-			display: none;
-		}
-
-		input[type="text"],
-		select {
-			border: none;
-			border-bottom: solid var(--black) 2px;
-			padding: 4px 0px;
-			font-family: var(--ui);
-			background-color: transparent;
-
-			&:focus {
-				outline: none;
-				border-bottom: var(--red) solid 2px;
-			}
-		}
-
-		input[type="range"] {
-			-webkit-appearance: none;
-			appearance: none;
-			background: transparent;
-			cursor: pointer;
-		}
-
-		input[type="number"] {
-			padding: 1px 5px;
-			width: 50px;
-			font-family: var(--ui);
-			background-color: transparent;
-			margin-left: 10px;
-
-			&:focus {
-				outline: none;
-				border-bottom: var(--red) solid 2px;
-			}
-		}
-
-		@supports (-moz-appearance: none) {
-			input[type="range"] {
-				appearance: auto;
-			}
-		}
-
-		$track-height: 5px;
-		$thumb-height: 15px;
-
-		input[type="range"]::-webkit-slider-runnable-track {
-			background: white;
-			border: solid 2px var(--black);
-			height: $track-height;
-			border-radius: 5px;
-			transform: translateY(13px);
-		}
-
-		input[type="range"]::-webkit-slider-thumb {
-			-webkit-appearance: none;
-			margin-top: calc(($track-height / 2) - ($thumb-height / 2) - 2px);
-			background: white;
-			border: var(--black) 2px solid;
-			border-radius: 50%;
-			height: $thumb-height;
-			width: 15px;
-
-			&:active {
-				background: var(--black);
-			}
-		}
-
-		textarea {
-			resize: none;
-			min-height: 105px;
-			font-size: 14px;
-			padding: 4px 6px;
-			border: solid var(--black) 2px;
-			font-family: var(--ui);
-
-			&:focus {
-				outline: none;
-				border: var(--red) solid 2px;
-			}
-		}
-
-		input[type="checkbox"] {
-			appearance: none;
-			cursor: pointer;
-			position: relative;
-
-			&::before {
-				content: "";
-				display: block;
-				height: 11px;
-				width: 11px;
-				border: 2px solid black;
-				box-shadow: 1px 1px 0px 0px var(--black), 2px 2px 0px 0px var(--black);
-				top: 0;
-				left: 0;
-			}
-
-			&:checked {
-				&::after {
-					content: "X";
-					font-weight: 800;
-					font-size: 12px;
-					position: absolute;
-					bottom: -1px;
-					left: 3px;
-				}
-			}
-		}
-	}
-}
