+++ src/components/Form.tsx
@@ -1,30 +1,29 @@
-import { useState } from "react";
-import Nav from "./form/Nav";
-import InterfaceTab from "./form/InterfaceTab";
-import UpgradesTab from "./form/UpgradesTab";
-import "./Form.scss";
+import { useState } from 'react';
+import Nav from './form/Nav';
+import InterfaceTab from './form/InterfaceTab';
+import UpgradesTab from './form/UpgradesTab';
 
 export default function Form({ updateAsset }) {
-	const [tab, setTab] = useState("interface");
+    const [tab, setTab] = useState('interface');
 
-	return (
-		<div className="form">
-			<form>
-				<Nav
-					tab={tab}
-					setTab={setTab}
-				/>
+    return (
+        <div className="form">
+            <form>
+                <Nav
+                    tab={tab}
+                    setTab={setTab}
+                />
 
-				<InterfaceTab
-					updateAsset={updateAsset}
-					tab={tab}
-				/>
+                <InterfaceTab
+                    updateAsset={updateAsset}
+                    tab={tab}
+                />
 
-				<UpgradesTab
-					updateAsset={updateAsset}
-					tab={tab}
-				/>
-			</form>
-		</div>
-	);
+                <UpgradesTab
+                    updateAsset={updateAsset}
+                    tab={tab}
+                />
+            </form>
+        </div>
+    );
 }
