--- src/components/Preview.scss
@@ -1,242 +1,0 @@
-.asset-preview.module {
-    .icon {
-        background-color: #d8cddd;
-    }
-    span.asset-type {
-        color: #7f5a8f;
-    }
-}
-
-.asset-preview.commandVehicle {
-    background: url(../assets/bg-colors/command-vehicle-1.svg) top -16px right 84px,
-        url(../assets/bg-colors/command-vehicle-2.svg) top 55px right -31px;
-
-    .icon {
-        background-color: #e0e3e6;
-    }
-    span.asset-type {
-        color: #9aa3ae;
-    }
-}
-
-.asset-preview.supportVehicle {
-    background: url(../assets/bg-colors/support-vehicle-1.svg) top -17px right 84px,
-        url(../assets/bg-colors/support-vehicle-2.svg) top 55px right -31px;
-
-    .icon {
-        background-color: #c8ccdd;
-    }
-    span.asset-type {
-        color: #4a5690;
-    }
-}
-
-.asset-preview.path {
-    background: url(../assets/bg-colors/path-1.svg) top -17px right 84px, url(../assets/bg-colors/path-2.svg) top 55px right -31px;
-
-    .icon {
-        background-color: #c4d8e5;
-    }
-    span.asset-type {
-        color: #417faa;
-    }
-}
-
-.asset-preview.companion {
-    background: url(../assets/bg-colors/companion-1.svg) top -17px right 84px, url(../assets/bg-colors/companion-2.svg) top 55px right -31px;
-
-    .icon {
-        background-color: #c4dcdc;
-    }
-    span.asset-type {
-        color: #3f8c89;
-    }
-}
-
-.asset-preview.deed {
-    background: url(../assets/bg-colors/deed-1.svg) top -17px right 84px, url(../assets/bg-colors/deed-2.svg) top 55px right -31px;
-
-    .icon {
-        background-color: #c4daca;
-    }
-    span.asset-type {
-        color: #4b974e;
-    }
-}
-
-.asset-preview {
-    & ~ button {
-        font-family: var(--ui);
-        width: 100%;
-        margin-top: 10px;
-        padding: 8px 0;
-        border-radius: 2px;
-        background: var(--black);
-        color: white;
-        border: none;
-
-        &:hover {
-            cursor: pointer;
-        }
-    }
-
-    height: 525px;
-    width: 375px;
-    border: var(--black) 2px solid;
-    padding: 15px;
-    background-repeat: no-repeat !important;
-    position: relative;
-    overflow: hidden;
-
-    header {
-        position: relative;
-        min-height: 80px;
-
-        .title-container {
-            display: flex;
-            flex-direction: column;
-            gap: 2px;
-
-            span.asset-type,
-            span.asset-title {
-                font-family: var(--title);
-                line-height: 85%;
-            }
-
-            span.asset-title {
-                font-weight: 900;
-                font-size: 26px;
-                max-width: 190px;
-                word-wrap: break-word;
-            }
-
-            span.asset-type {
-                font-size: 15px;
-                font-weight: 700;
-            }
-
-            span.asset-description {
-                font-family: var(--text);
-                font-size: 18px;
-                word-wrap: break-word;
-                padding-right: 85px;
-                margin-bottom: 10px;
-            }
-
-            span.asset-additional-field {
-                border-bottom: solid var(--black) 2px;
-                font-family: var(--title);
-                font-size: 12px;
-                width: 250px;
-                font-weight: 700;
-                color: #cbcfd4;
-                word-wrap: break-word;
-                margin-bottom: 6px;
-            }
-        }
-
-        .clip {
-            clip-path: url(#svg-hex);
-            overflow: hidden;
-        }
-
-        .icon-container {
-            $height: 80px;
-            $width: 75px;
-            $gap: 5px;
-            display: flex;
-            justify-content: center;
-            align-items: center;
-            background-color: var(--black);
-            height: $height;
-            width: $width;
-            position: absolute;
-            top: 0px;
-            right: 5px;
-
-            .icon {
-                height: $height - $gap;
-                width: $width - $gap;
-                display: flex;
-                justify-content: center;
-                align-items: center;
-            }
-        }
-    }
-
-    ul {
-        display: flex;
-        flex-direction: column;
-        gap: 4px;
-
-        li {
-            list-style: none;
-            font-family: var(--text);
-            font-size: 18px;
-            white-space: pre-line;
-            padding-left: 25px;
-            margin-right: 10px;
-            background-position: 0 2px;
-            background-repeat: no-repeat;
-            background-image: url(../assets/hex-border.svg);
-
-            &.checked {
-                background-image: url(../assets/hex-filled.svg);
-            }
-
-            b,
-            strong {
-                font-weight: bold;
-            }
-
-            i,
-            em {
-                font-style: italic;
-            }
-        }
-    }
-
-    .asset-track {
-        display: flex;
-        position: absolute;
-        bottom: 15px;
-        .asset-track-hex {
-            background-image: url(../assets/hex-border-track.svg);
-            background-repeat: no-repeat;
-            background-size: cover;
-            height: 38px;
-            width: 34px;
-            text-align: center;
-            line-height: 40px;
-            font-family: var(--title);
-            font-size: 14px;
-            color: #c7cacf;
-            font-weight: 700;
-            margin-left: -5px;
-        }
-    }
-}
-
-.top-hex {
-    position: absolute;
-    top: -32px;
-    right: 70px;
-}
-
-.bottom-hex {
-    position: absolute;
-    top: 40px;
-    right: -46px;
-}
-
-.color-grid {
-    display: grid;
-    grid-template-columns: repeat(25, 1fr);
-    gap: 5px;
-
-    div {
-        width: 12px;
-        aspect-ratio: 1 / 1;
-        cursor: pointer;
-    }
-}
