+++ src/components/preview/AditionnalField.tsx
@@ -1,0 +1,9 @@
+export default function AdditionalField({ id, text }: { id: number; text: string }) {
+    return (
+        <span
+            className="asset-additional-field"
+            key={id}>
+            {text}
+        </span>
+    );
+}
