+++ src/main.jsx
@@ -1,10 +1,10 @@
-import React from 'react'
-import ReactDOM from 'react-dom/client'
-import App from './App.jsx'
-import './index.scss'
+import React from 'react';
+import ReactDOM from 'react-dom/client';
+import App from 'src/App.jsx';
+import 'src/index.scss';
 
 ReactDOM.createRoot(document.getElementById('root')).render(
-  <React.StrictMode>
-    <App/>
-  </React.StrictMode>,
-)
+    <React.StrictMode>
+        <App />
+    </React.StrictMode>
+);
